+++ main.py
@@ -1,13 +1,10 @@
-users: list = [
-    {"name": "Bernard","location":"Etk","posts":400},
-    {"name": "Krzysztof","location":"Białobrzegi","posts":500},
-    {"name": "Maja","location":"Świecie", "posts":300},
-    {"name": "Zuzanna","location":"Radzyń_Podlaski","posts":700},
-]
+from utils.model import users
+from utils.controller import get_user_info
 
-def get_user_info(user_data:list)->None:
 
-    for user in user_data:
-        print(f"Twój znajomy {user['name']}! z mieiscowosci {user['location']} opublikował {user[ 'posts']} postów")
+def main():
+    get_user_info(users)
 
-get_user_info(users)
+
+if __name__ == "__main__":
+    main()
