+++ utils/model.py
@@ -1,0 +1,6 @@
+users: list = [
+    {"name": "Bernard","location":"Etk","posts":400},
+    {"name": "Krzysztof","location":"Białobrzegi","posts":500},
+    {"name": "Maja","location":"Świecie", "posts":300},
+    {"name": "Zuzanna","location":"Radzyń_Podlaski","posts":700},
+]
